feat: serial VI IDs, delay builtins, and smarter advertise IP
Replace UUID template IDs with BIGSERIAL, drop current-agent holder,
and add kind=delay for Agent general-page sleep steps in sequences.
Prefer LAN egress IP toward center over Mihomo 198.18 fake-ip.

+++ crates/scheduler/migrations/001_init.sql
@@ -2,11 +2,11 @@ CREATE TABLE IF NOT EXISTS agents (
   id TEXT PRIMARY KEY,
   name TEXT NOT NULL,
   ip TEXT NOT NULL,
-  port INTEGER NOT NULL,
+  port BIGINT NOT NULL,
   status TEXT NOT NULL,
   cpu_percent REAL NOT NULL DEFAULT 0,
   memory_percent REAL NOT NULL DEFAULT 0,
-  busy INTEGER NOT NULL DEFAULT 0,
+  busy BIGINT NOT NULL DEFAULT 0,
   last_seen_at TEXT,
   created_at TEXT NOT NULL,
   UNIQUE(name, ip, port)
@@ -18,7 +18,7 @@ CREATE TABLE IF NOT EXISTS task_templates (
   shell TEXT NOT NULL,
   command TEXT NOT NULL,
   workdir TEXT,
-  timeout_secs INTEGER NOT NULL,
+  timeout_secs BIGINT NOT NULL,
   created_at TEXT NOT NULL
 );
 
@@ -30,7 +30,7 @@ CREATE TABLE IF NOT EXISTS tasks (
   shell TEXT NOT NULL,
   command TEXT NOT NULL,
   workdir TEXT,
-  timeout_secs INTEGER NOT NULL,
+  timeout_secs BIGINT NOT NULL,
   status TEXT NOT NULL,
   exit_code INTEGER,
   stdout TEXT NOT NULL DEFAULT '',
